Capture code block output before printing it out
When a code block invokes methods from the Test module, like is(), ok(),
and such, if some fail, output goes to STDERR, and sometimes that error
output is printed only after the label of the next block to run is
printed. Not sure what's going on exactly, but capturing like with the
code from this commit appears to fix it.